Move class to common module for easy reuse
